requirements.txt
numpy==1.24.2 
pandas==1.5.3
pyarrow==11.0.0
scikit-learn==1.2.1
jupyter==1.0.0
notebook==6.5.2
matplotlib==3.7.0
tensorboard==2.12.0

torch==1.13.1
torchvision==0.14.1
transformers==4.26.1
pytorch-metric-learning==2.0.0
open-clip-torch==2.14.0
sentence_transformers==2.2.2
faiss-gpu==1.7.2

kaggle==1.5.12
# hydra-core==1.3.1
# hydra-colorlog==1.2.0